Agent1-DocReader AI Agent/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
python-dotenv==1.0.0
